Moved test.py to tests folder.
Conventions say that test code should not be part of the package.
I edited a few files accordingly.
Normally tests folder does not have __init__, but this is a pragmatic
fix in order to make things work with older Python versions like 2.7,
that don't know about folders as namespaces.

Everything works, and all tests pass.